+++ test/DependencyInjection/SentryExtensionTest.php
@@ -13,6 +13,7 @@ use Sentry\Monolog\Handler;
 use Sentry\Options;
 use Sentry\SentryBundle\DependencyInjection\SentryExtension;
 use Sentry\SentryBundle\EventListener\ErrorListener;
+use Sentry\SentryBundle\EventListener\MessengerListener;
 use Sentry\SentryBundle\Test\BaseTestCase;
 use Sentry\SentrySdk;
 use Sentry\Severity;
@@ -61,7 +62,7 @@ class SentryExtensionTest extends BaseTestCase
 
         $this->assertNull($options->getDsn());
         $this->assertSame('test', $options->getEnvironment());
-        $this->assertSame([realpath('./var/cache'), '/dir/project/root/vendor'], $options->getInAppExcludedPaths());
+        $this->assertSame([$container->getParameter('kernel.cache_dir'), '/dir/project/root/vendor'], $options->getInAppExcludedPaths());
 
         $this->assertSame(1, $container->getParameter('sentry.listener_priorities.request'));
         $this->assertSame(1, $container->getParameter('sentry.listener_priorities.sub_request'));
@@ -411,6 +412,17 @@ class SentryExtensionTest extends BaseTestCase
         ]);
 
         $this->assertFalse($container->has(self::MONOLOG_HANDLER_TEST_PUBLIC_ALIAS));
+    }
+
+    public function testMessengerHandlerIsNotRegistered(): void
+    {
+        $container = $this->getContainer([
+            'messenger' => [
+                'enabled' => false,
+            ],
+        ]);
+
+        $this->assertFalse($container->has(MessengerListener::class));
     }
 
     private function getContainer(array $configuration = []): Container